Wire up account manager and login window
On app startup, we retrieve the list of account information. If there are no accounts, a login window is invoked.

Login windows trigger an account list update upon their completion. The app watches for account list updates, and invokes the main window whenever there's a new account.

The logout functionality no longer clears local storage and quits the application, but instead it clears local storage, closes all of the open windows, and resets the app to its initial state.