Soften navigation restrictions around login
There are a ton of reasonable accounts.google.com URLs which should be allowed to support Google's wide variety of login flows. A very specific navigation allowance list likely won't be feasible for that domain. This opens it up, but excludes TOS pages specifically to force those guys to open in a new window.